requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.8.7
django-allauth==0.32.0
django-ckeditor==5.4.0
django-countries==5.1.1
django-fontawesome==0.3.1
django-jet==1.0.7
django-js-asset==1.0.0
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
oauthlib==2.0.6
Pillow==5.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-openid==2.2.5
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
sorl-thumbnail==12.4.1
uritemplate==3.0.0
urllib3==1.22
